chore: Move more task tests to account level (#3971)
Move the remaining tests checking the default parameter values.
