+++ pkg/testacc/resource_task_acceptance_test.go
@@ -19,7 +19,6 @@ import (
 	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/acceptance/bettertestspoc/config"
 	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/acceptance/bettertestspoc/config/model"
 	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/acceptance/helpers/random"
-	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/helpers"
 	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/provider/resources"
 	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/sdk"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
@@ -28,193 +27,6 @@ import (
 )
 
 // TODO(SNOW-1822118): Create more complicated tests for task
-
-func TestAcc_Task_Basic(t *testing.T) {
-	currentRole := testClient().Context.CurrentRole(t)
-
-	id := testClient().Ids.RandomSchemaObjectIdentifier()
-	statement := "SELECT 1"
-
-	configModel := model.TaskWithId("test", id, false, statement)
-
-	resource.Test(t, resource.TestCase{
-		ProtoV6ProviderFactories: TestAccProtoV6ProviderFactories,
-		TerraformVersionChecks: []tfversion.TerraformVersionCheck{
-			tfversion.RequireAbove(tfversion.Version1_5_0),
-		},
-		CheckDestroy: CheckDestroy(t, resources.Task),
-		Steps: []resource.TestStep{
-			{
-				Config: config.FromModels(t, configModel),
-				Check: assertThat(t,
-					resourceassert.TaskResource(t, configModel.ResourceReference()).
-						HasFullyQualifiedNameString(id.FullyQualifiedName()).
-						HasDatabaseString(id.DatabaseName()).
-						HasSchemaString(id.SchemaName()).
-						HasNameString(id.Name()).
-						HasStartedString(r.BooleanFalse).
-						HasWarehouseString("").
-						HasNoScheduleSet().
-						HasConfigString("").
-						HasAllowOverlappingExecutionString(r.BooleanDefault).
-						HasErrorIntegrationString("").
-						HasCommentString("").
-						HasFinalizeString("").
-						HasAfter().
-						HasWhenString("").
-						HasSqlStatementString(statement),
-					resourceshowoutputassert.TaskShowOutput(t, configModel.ResourceReference()).
-						HasCreatedOnNotEmpty().
-						HasName(id.Name()).
-						HasIdNotEmpty().
-						HasDatabaseName(id.DatabaseName()).
-						HasSchemaName(id.SchemaName()).
-						HasOwner(currentRole.Name()).
-						HasComment("").
-						HasWarehouse(sdk.NewAccountObjectIdentifier("")).
-						HasScheduleEmpty().
-						HasPredecessors().
-						HasState(sdk.TaskStateSuspended).
-						HasDefinition(statement).
-						HasCondition("").
-						HasAllowOverlappingExecution(false).
-						HasErrorIntegration(sdk.NewAccountObjectIdentifier("")).
-						HasLastCommittedOn("").
-						HasLastSuspendedOn("").
-						HasOwnerRoleType("ROLE").
-						HasConfig("").
-						HasBudget("").
-						HasTaskRelations(sdk.TaskRelations{}),
-					resourceparametersassert.TaskResourceParameters(t, configModel.ResourceReference()).
-						HasAllDefaults(),
-				),
-			},
-			{
-				ResourceName: configModel.ResourceReference(),
-				ImportState:  true,
-				ImportStateCheck: assertThatImport(t,
-					resourceassert.ImportedTaskResource(t, helpers.EncodeResourceIdentifier(id)).
-						HasFullyQualifiedNameString(id.FullyQualifiedName()).
-						HasDatabaseString(id.DatabaseName()).
-						HasSchemaString(id.SchemaName()).
-						HasNameString(id.Name()).
-						HasStartedString(r.BooleanFalse).
-						HasWarehouseString("").
-						HasNoScheduleSet().
-						HasConfigString("").
-						HasAllowOverlappingExecutionString(r.BooleanFalse).
-						HasErrorIntegrationString("").
-						HasCommentString("").
-						HasFinalizeString("").
-						HasAfterEmpty().
-						HasWhenString("").
-						HasSqlStatementString(statement),
-				),
-			},
-		},
-	})
-}
-
-func TestAcc_Task_Complete(t *testing.T) {
-	currentRole := testClient().Context.CurrentRole(t)
-
-	errorNotificationIntegration, errorNotificationIntegrationCleanup := testClient().NotificationIntegration.CreateWithGcpPubSub(t)
-	t.Cleanup(errorNotificationIntegrationCleanup)
-
-	id := testClient().Ids.RandomSchemaObjectIdentifier()
-	statement := "SELECT 1"
-	taskConfig := `{"output_dir": "/temp/test_directory/", "learning_rate": 0.1}`
-	comment := random.Comment()
-	condition := `SYSTEM$STREAM_HAS_DATA('MYSTREAM')`
-	configModel := model.TaskWithId("test", id, true, statement).
-		WithWarehouse(testClient().Ids.WarehouseId().Name()).
-		WithScheduleMinutes(10).
-		WithConfigValue(configvariable.StringVariable(taskConfig)).
-		WithAllowOverlappingExecution(r.BooleanTrue).
-		WithErrorIntegration(errorNotificationIntegration.ID().Name()).
-		WithComment(comment).
-		WithWhen(condition)
-
-	resource.Test(t, resource.TestCase{
-		ProtoV6ProviderFactories: TestAccProtoV6ProviderFactories,
-		TerraformVersionChecks: []tfversion.TerraformVersionCheck{
-			tfversion.RequireAbove(tfversion.Version1_5_0),
-		},
-		CheckDestroy: CheckDestroy(t, resources.Task),
-		Steps: []resource.TestStep{
-			{
-				ConfigDirectory: ConfigurationDirectory("TestAcc_Task/basic"),
-				ConfigVariables: config.ConfigVariablesFromModel(t, configModel),
-				Check: assertThat(t,
-					resourceassert.TaskResource(t, configModel.ResourceReference()).
-						HasFullyQualifiedNameString(id.FullyQualifiedName()).
-						HasDatabaseString(id.DatabaseName()).
-						HasSchemaString(id.SchemaName()).
-						HasNameString(id.Name()).
-						HasStartedString(r.BooleanTrue).
-						HasWarehouseString(testClient().Ids.WarehouseId().Name()).
-						HasScheduleMinutes(10).
-						HasConfigString(taskConfig).
-						HasAllowOverlappingExecutionString(r.BooleanTrue).
-						HasErrorIntegrationString(errorNotificationIntegration.ID().Name()).
-						HasCommentString(comment).
-						HasFinalizeString("").
-						HasAfterEmpty().
-						HasWhenString(condition).
-						HasSqlStatementString(statement),
-					resourceshowoutputassert.TaskShowOutput(t, configModel.ResourceReference()).
-						HasCreatedOnNotEmpty().
-						HasName(id.Name()).
-						HasIdNotEmpty().
-						HasDatabaseName(id.DatabaseName()).
-						HasSchemaName(id.SchemaName()).
-						HasOwner(currentRole.Name()).
-						HasComment(comment).
-						HasWarehouse(testClient().Ids.WarehouseId()).
-						HasScheduleMinutes(10).
-						HasPredecessors().
-						HasState(sdk.TaskStateStarted).
-						HasDefinition(statement).
-						HasCondition(condition).
-						HasAllowOverlappingExecution(true).
-						HasErrorIntegration(errorNotificationIntegration.ID()).
-						HasLastCommittedOnNotEmpty().
-						HasLastSuspendedOn("").
-						HasOwnerRoleType("ROLE").
-						HasConfig(taskConfig).
-						HasBudget("").
-						HasTaskRelations(sdk.TaskRelations{}),
-					resourceparametersassert.TaskResourceParameters(t, configModel.ResourceReference()).
-						HasAllDefaults(),
-				),
-			},
-			{
-				ResourceName:    configModel.ResourceReference(),
-				ImportState:     true,
-				ConfigDirectory: ConfigurationDirectory("TestAcc_Task/basic"),
-				ConfigVariables: config.ConfigVariablesFromModel(t, configModel),
-				ImportStateCheck: assertThatImport(t,
-					resourceassert.ImportedTaskResource(t, helpers.EncodeResourceIdentifier(id)).
-						HasFullyQualifiedNameString(id.FullyQualifiedName()).
-						HasDatabaseString(id.DatabaseName()).
-						HasSchemaString(id.SchemaName()).
-						HasNameString(id.Name()).
-						HasStartedString(r.BooleanTrue).
-						HasWarehouseString(testClient().Ids.WarehouseId().Name()).
-						HasScheduleMinutes(10).
-						HasConfigString(taskConfig).
-						HasAllowOverlappingExecutionString(r.BooleanTrue).
-						HasErrorIntegrationString(errorNotificationIntegration.ID().Name()).
-						HasCommentString(comment).
-						HasFinalizeString("").
-						HasAfterEmpty().
-						HasWhenString(condition).
-						HasSqlStatementString(statement),
-				),
-			},
-		},
-	})
-}
 
 func TestAcc_Task_Updates(t *testing.T) {
 	currentRole := testClient().Context.CurrentRole(t)
